get the base shape working
